+++ src/components/styles/_variables.scss
@@ -9,4 +9,7 @@ $offwhite:#ece8d3;
 $font-stack-primary: 'Lato', sans-serif;;
 $font-stack-secondary: 'Lato', serif;
 
+//Box Shadow
+$shadow: 0 8px 16px 0 rgba(0,0,0,0.2), 0 6px 10px 0 rgba(0,0,0,0.19);
+
