+++ src/components/styles/signedup.module.scss
@@ -1,0 +1,60 @@
+@import '../styles/variables';
+
+@media screen and (min-width:320px) {
+  .signup {
+    background-color: #eee;
+    display: grid;
+  }
+
+  .container {
+    padding: 1em;
+    
+    
+    .form {
+    border-radius: 6px;
+    box-shadow: $shadow;
+    background-color: rgb(191, 175, 156);
+    display: grid;
+    grid-template-columns: 80%;
+    justify-content: center;
+    align-items: stretch;
+    }
+
+    .label {
+      margin-top: 1em;
+
+    }
+        
+    .input {
+      border-radius: 6px;
+      margin-top: 1em;
+      border: none;
+      padding: 1em;
+    }
+    
+    ul {
+      padding: 0;
+      list-style-type: none;
+      display: unset;
+    }
+
+    button {
+      color: #eee;
+      border-radius: 6px;
+      box-shadow: $shadow;
+      font-weight: bold;
+      font-size: 16px;
+      margin-bottom: 5%;
+      border: none;
+      background-color: #006344;
+      padding: .7em 40% .7em 36%;
+    }
+  }
+}
+
+// @media screen and (min-width: 768px) {
+//   button {
+//     grid-template-columns: 50%;
+//   }
+// }
+
